Set the label for the token upload button based on the state of the token file
This is so that setting up the widget twice (which might happen
because clicking the upload button sets the current cell back to the one
that set up the button) won't cause the label to change back to
"Please upload a token".  I haven't found a way to have clicking the button
advance the current cell yet.